chatbot-ia-generative/chat/services/multi_search.py
import requests
from typing import List, Dict
import logging
import json
from bs4 import BeautifulSoup
import time
from django.conf import settings

logger = logging.getLogger(__name__)

# Configure colored logs
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)


class MultiSearchService:
    """Robust search service that tries multiple methods."""
    
    def __init__(self):
        self.max_results = settings.MAX_SEARCH_RESULTS
    
    def search(self, query: str) -> List[Dict]:
        """
        Try multiple search methods in order of preference.
        """
        logger.info(f"\n🌍 RECHERCHE WEB MULTI-SOURCE")
        logger.info(f"🔍 Query: '{query}'")
        
        # Method 1: Try Google search scraping (most reliable)
        logger.info(f"\n1️⃣ Tentative Google Search...")
        results = self._google_search_scrape(query)
        if results:
            logger.info(f"✅ Google Search réussi: {len(results)} résultats")
            return results
        logger.warning(f"❌ Google Search échoué")
        
        # Method 2: Try Bing search scraping
        logger.info(f"\n2️⃣ Tentative Bing Search...")
        results = self._bing_search_scrape(query)
        if results:
            logger.info(f"✅ Bing Search réussi: {len(results)} résultats")
            return results
        logger.warning(f"❌ Bing Search échoué")
        
        # Method 3: Try direct news sites
        logger.info(f"\n3️⃣ Tentative sites d'actualités directs...")
        results = self._direct_news_search(query)
        if results:
            logger.info(f"✅ Recherche directe réussie: {len(results)} résultats")
            return results
        logger.warning(f"❌ Recherche directe échouée")
        
        # If all fail, return mock data for demo
        logger.warning(f"\n⚠️ TOUTES LES MÉTHODES ONT ÉCHOUÉ - Utilisation des données de démo")
        return self._get_demo_results(query)
    
    def _google_search_scrape(self, query: str) -> List[Dict]:
        """Scrape Google search results."""
        try:
            headers = {
                'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
            }
            
            params = {
                'q': query + ' latest news 2025',
                'tbm': 'nws',  # News search
                'tbs': 'qdr:w'  # Last week
            }
            
            logger.info(f"   🌐 Requête vers Google Search...")
            response = requests.get(
                'https://www.google.com/search',
                headers=headers,
                params=params,
                timeout=5
            )
            logger.info(f"   📡 Status code: {response.status_code}")
            
            if response.status_code == 200:
                soup = BeautifulSoup(response.text, 'html.parser')
                results = []
                
                # Find news articles
                articles = soup.find_all('div', {'class': 'SoaBEf'}) or soup.find_all('article')
                
                for article in articles[:self.max_results]:
                    title_elem = article.find('h3') or article.find('a')
                    link_elem = article.find('a', href=True)
                    snippet_elem = article.find('div', {'class': 'GI74Re'}) or article.find('span')
                    
                    if title_elem and link_elem:
                        results.append({
                            'title': title_elem.get_text(strip=True),
                            'url': link_elem['href'] if link_elem['href'].startswith('http') else f"https://google.com{link_elem['href']}",
                            'content': snippet_elem.get_text(strip=True) if snippet_elem else ''
                        })
                
                if results:
                    logger.info(f"   ✅ Extraction réussie: {len(results)} articles trouvés")
                    for i, r in enumerate(results[:3]):
                        logger.info(f"     {i+1}. {r['title'][:60]}...")
                    return results
                else:
                    logger.warning(f"   ⚠️ Aucun article extrait du HTML")
                    
        except Exception as e:
            logger.error(f"   ❌ Erreur Google scrape: {str(e)}")
        
        return []
    
    def _bing_search_scrape(self, query: str) -> List[Dict]:
        """Scrape Bing search results."""
        try:
            headers = {
                'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
            }
            
            params = {
                'q': query + ' latest AI news 2025',
                'filters': 'ex1:"ez1"'  # Recent results
            }
            
            response = requests.get(
                'https://www.bing.com/news/search',
                headers=headers,
                params=params,
                timeout=5
            )
            
            if response.status_code == 200:
                soup = BeautifulSoup(response.text, 'html.parser')
                results = []
                
                # Find news cards
                news_cards = soup.find_all('div', {'class': 'news-card'})
                
                for card in news_cards[:self.max_results]:
                    title_elem = card.find('a', {'class': 'title'})
                    snippet_elem = card.find('div', {'class': 'snippet'})
                    
                    if title_elem:
                        results.append({
                            'title': title_elem.get_text(strip=True),
                            'url': title_elem.get('href', ''),
                            'content': snippet_elem.get_text(strip=True) if snippet_elem else ''
                        })
                
                if results:
                    logger.info(f"   ✅ Bing: {len(results)} résultats trouvés")
                    return results
                    
        except Exception as e:
            logger.error(f"   ❌ Erreur Bing: {str(e)}")
        
        return []
    
    def _direct_news_search(self, query: str) -> List[Dict]:
        """Search directly on news sites."""
        results = []
        logger.info(f"   📰 Recherche sur sites tech spécialisés...")
        
        # Tech news sites that often have AI news
        news_sites = [
            {
                'name': 'TechCrunch AI',
                'url': 'https://techcrunch.com/category/artificial-intelligence/',
                'selector': 'article'
            },
            {
                'name': 'The Verge AI',
                'url': 'https://www.theverge.com/ai-artificial-intelligence',
                'selector': 'article'
            }
        ]
        
        for site in news_sites:
            try:
                response = requests.get(site['url'], timeout=3)
                if response.status_code == 200:
                    soup = BeautifulSoup(response.text, 'html.parser')
                    articles = soup.find_all(site['selector'], limit=2)
                    
                    for article in articles:
                        title = article.find('h2') or article.find('h3')
                        link = article.find('a', href=True)
                        
                        if title and link:
                            results.append({
                                'title': title.get_text(strip=True),
                                'url': link['href'] if link['href'].startswith('http') else site['url'] + link['href'],
                                'content': f"From {site['name']}: Latest AI news and developments"
                            })
                            
            except Exception as e:
                logger.error(f"Direct news search error for {site['name']}: {str(e)}")
        
        return results[:self.max_results]
    
    def _get_demo_results(self, query: str) -> List[Dict]:
        """Return demo results for testing."""
        logger.info(f"   🎭 Génération de résultats de démonstration")
        if 'ia' in query.lower() or 'ai' in query.lower():
            return [
                {
                    'title': 'OpenAI Announces GPT-5 with Advanced Reasoning Capabilities',
                    'url': 'https://openai.com/blog/gpt-5-announcement',
                    'content': 'OpenAI has unveiled GPT-5, featuring breakthrough improvements in logical reasoning, mathematical problem-solving, and multi-modal understanding. The model demonstrates near-human performance on complex reasoning tasks.'
                },
                {
                    'title': 'Google DeepMind Introduces Gemini 2.0 Ultra',
                    'url': 'https://deepmind.google/gemini-2-ultra',
                    'content': 'Google DeepMind announces Gemini 2.0 Ultra, a multimodal AI model that seamlessly processes text, images, audio, and video with state-of-the-art performance across all modalities.'
                },
                {
                    'title': 'Anthropic Releases Claude 3.5 with 1M Token Context',
                    'url': 'https://anthropic.com/claude-3-5',
                    'content': 'Anthropic has launched Claude 3.5, featuring an unprecedented 1 million token context window, enabling analysis of entire codebases, books, and extensive documents in a single conversation.'
                }
            ]
        
        return [{
            'title': f'Search result for: {query}',
            'url': 'https://example.com',
            'content': 'Demo search result content'
        }]